+++ docs/history.rst
@@ -1,6 +1,10 @@
 History
 =======
 
+Latest
+------
+
+
 0.7.1
 ------
 - BUG: Handle transforms with rotation (pull #401)